pyproject.toml
[tool.poetry]
name = "thsr-helper"
version = "0.1.0"
description = ""
authors = ["Nick Kuo <nickest14@gmail.com>"]
readme = "README.md"
packages = [{include = "thsr_helper"}]

[tool.poetry.dependencies]
python = "3.12.2"
typer = "^0.9.0"
rich = "^13.7.0"
pre-commit = "^3.6.1"
tomli = "^2.0.1"
tomlkit = "^0.12.3"
requests = "^2.31.0"
bs4 = "^0.0.2"
pillow = "^10.2.0"
pydantic = "^2.6.1"
tinydb = "^4.8.0"
pytz = "^2024.1"

[tool.poetry.scripts]
thsr-helper = "thsr_helper.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
